Fix rule interpretations to match modern Derby (#2)
Playing Derby follows a variety of rule interpretations and house-rules, as the rulebook is not very explicit about interpretations. See Readme for details.

However, after acquiring an edition of the modern Derby game, that also has a classic edition, the rules for classic Derby are explicitly written in their rule book. Although players can still choose their own house rules or rule interpretations when playing the original game, I think it makes sense to have some consistency and adhere to the modern interpretation, even for the original game.

* Changed interpretation of SteadyPace to pick up gallop card during home stretch. This is explicitly stated in the new rulebook, and is retrofitted the Drechsler edition.

* Fixing logic for await, which should 1) await the nearest (not specific horse), and 2) await the last (not specific horse).

* Fixed critical bug, where if a horse is eliminated and is the last horse, the turn order is increment incorrectly, and skips the turn of the first horse.